+++ src/icons/icon-success.tsx
@@ -28,26 +28,35 @@ const IconSuccessO2: React.FC<Props> = ({size = 64, color, skipAnimation = false
         <svg role="presentation" width={size} height={size} viewBox="0 0 64 64">
             <g
                 className={classnames({[classes.iconContainer]: !color})}
-                strokeWidth="2"
                 strokeLinecap="round"
                 strokeLinejoin="round"
             >
                 <path
                     fill="none"
+                    strokeWidth="2"
                     stroke={color ? color : undefined}
-                    strokeMiterlimit="10"
-                    d="M21.7,35.6l2.8,2.8c1.1,1.1,2.9,1.1,4.1,0l17.3-17.3"
-                    {...(!skipAnimation &&
-                        mergeProperties(
-                            getAnimateDrawLineProps('33', '0.9s', platformOverrides),
-                            getAnimateHopInProps('0.9s', platformOverrides)
-                        ))}
+                    d="M45.64 53.84A25.599 25.599 0 0 1 32 57.75C17.78 57.75 6.25 46.22 6.25 32S17.78 6.25 32 6.25 57.75 17.78 57.75 32c0 5.13-1.5 9.92-4.09 13.93"
+                    {...(!skipAnimation && getAnimateDrawLineProps('202', '0.3s', platformOverrides))}
                 />
                 <path
                     fill="none"
+                    strokeWidth="2"
                     stroke={color ? color : undefined}
-                    d="M49.1,57.9C44.2,61.1,38.3,63,32,63C14.9,63,1,49.1,1,32C1,14.9,14.9,1,32,1c17.1,0,31,13.9,31,31c0,7.5-2.6,14.3-7.1,19.7"
-                    {...(!skipAnimation && getAnimateDrawLineProps('186', '0.5s', platformOverrides))}
+                    d="m19.33 30.39 9.88 10.95 18.13-19.23"
+                    {...(!skipAnimation &&
+                        mergeProperties(
+                            getAnimateDrawLineProps('44', '0.6s', platformOverrides),
+                            getAnimateHopInProps('0.6', platformOverrides)
+                        ))}
+                />
+                <circle
+                    fill="none"
+                    strokeWidth="2"
+                    stroke={color ? color : undefined}
+                    cx="50.18"
+                    cy="50.1"
+                    r=".89"
+                    {...(!skipAnimation && getAnimateDrawLineProps('180', '0.2s', platformOverrides))}
                 />
             </g>
         </svg>