+++ src/index.tsx
@@ -42,6 +42,7 @@ export {default as MasterDetailLayout} from './master-detail-layout';
 export {default as NavigationBreadcrumbs} from './navigation-breadcrumbs';
 export {default as Text, Text1, Text2, Text3, Text4, Text5, Text6, Text7, Text8, Text9, Text10} from './text';
 export {default as Tag} from './tag';
+export type {TagType} from './tag';
 export {default as SectionTitle} from './section-title';
 export {Placeholder, AvatarPlaceholder} from './placeholder';
 export {RowList, Row, BoxedRowList, BoxedRow} from './list';
@@ -71,11 +72,13 @@ export {
 } from './navigation-bar';
 export {default as Image} from './image';
 export {default as Chip} from './chip';
+export {default as Video} from './video';
 
 // Forms
 export {default as Form} from './form';
 export {default as Select} from './select';
 export {default as TextField} from './text-field';
+export {TextFieldBase} from './text-field-base';
 export {default as SearchField} from './search-field';
 export {default as EmailField} from './email-field';
 export {default as PhoneNumberField} from './phone-number-field';